+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## Unreleased
+
+### Added
+
+- expose API for transport validation
+
 ## 0.7.25 2020-04-12
 
 ### Fixed